backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
langchain-google-genai
langchain-community
faiss-cpu
pypdf
beautifulsoup4
python-dotenv
notion-client